feat: export games and protocols besides GameDig
